+++ cmd/glaze/main.go
@@ -8,6 +8,8 @@ import (
 	"github.com/go-go-golems/glazed/pkg/doc"
 	"github.com/go-go-golems/glazed/pkg/help"
 	help_cmd "github.com/go-go-golems/glazed/pkg/help/cmd"
+	"github.com/go-go-golems/glazed/pkg/help/server"
+	"github.com/go-go-golems/glazed/pkg/web"
 	"github.com/spf13/cobra"
 )
 
@@ -30,8 +32,10 @@ func main() {
 	err = doc.AddDocToHelpSystem(helpSystem)
 	cobra.CheckErr(err)
 
-	// Set up help system with UI support
 	help_cmd.SetupCobraRootCommand(helpSystem, rootCmd)
+
+	// Add the serve command (embeds the React SPA via pkg/web).
+	rootCmd.AddCommand(server.NewServeCommand(helpSystem, web.FS))
 
 	// JSON command
 	jsonCmd, err := cmds.NewJsonCommand()